requirements.txt
flask
torch
torchvision
opencv-python
numpy
Pillow
ultralytics